refactor: 사용하지않은 코드 삭제, 컨벤션 맞지않던 코드 수정, 크게 의미없는 코드 수정
