--- src/main/java/attendance/domain/dto/AttendanceDto.java
@@ -1,6 +1,0 @@
-package attendance.domain.dto;
-
-import java.time.LocalDateTime;
-
-public record AttendanceDto(String nickname, LocalDateTime attendanceDate) {
-}
