--- src/main/java/attendance/domain/dto/AttendanceHistoryDto.java
@@ -1,4 +1,0 @@
-package attendance.domain.dto;
-
-public record AttendanceHistoryDto(String nickname) {
-}
