--- src/main/java/attendance/domain/dto/AttendanceModifyDto.java
@@ -1,6 +1,0 @@
-package attendance.domain.dto;
-
-import java.time.LocalDateTime;
-
-public record AttendanceModifyDto(String nickname, LocalDateTime attendanceDate) {
-}
