--- src/main/java/attendance/domain/dto/CrewAttendanceHistory.java
@@ -1,4 +1,0 @@
-package attendance.domain.dto;
-
-public record CrewAttendanceHistory(String nickname, attendance.domain.AttendanceHistory attendanceHistory) {
-}
