--- src/main/java/attendance/domain/dto/CrewDismissDto.java
@@ -1,7 +1,0 @@
-package attendance.domain.dto;
-
-import attendance.domain.AttendanceHistory;
-
-public record CrewDismissDto(String nickname, AttendanceHistory attendanceHistory) {
-
-}
