--- src/main/java/attendance/domain/dto/DateValidateDto.java
@@ -1,4 +1,0 @@
-package attendance.domain.dto;
-
-public record DateValidateDto(java.time.LocalDate date) {
-}
