--- src/main/java/attendance/domain/dto/NicknameValidateDto.java
@@ -1,4 +1,0 @@
-package attendance.domain.dto;
-
-public record NicknameValidateDto(String nickname) {
-}
