--- src/main/java/attendance/domain/dto/TimeValidateDto.java
@@ -1,4 +1,0 @@
-package attendance.domain.dto;
-
-public record TimeValidateDto(java.time.LocalTime time) {
-}
